Update for Drawpile 2.2.0
Changes around the setup script to not be just for an old Debian version
anymore, instead being compatible with current Debian, Ubuntu and Arch.
It just uses docker and docker-compose fro m the repos now, since
they're fine for our purposes. The webadmin is grabbed as a pre-built
version instead of wasting time building it by hand.