Work around the docker compose/docker-compose mess
On Ubuntu, docker compose v1 and v2 are in different packages and use
different commands to invoke them. Older Ubuntus don't have the latter,
newer ones can't use v1 anymore because it crashes on startup. So we now
have to juggle around that by conditionally installing stuff based on
Ubuntu version and having a wrapper script that invokes the correct
version.